Bug fixes in chat friendlist and chatlist queries + change in transactions schema details

+++ database/schema.sql
@@ -49,7 +49,7 @@ CREATE TABLE IF NOT EXISTS transactions (
   type trade_type,
   datetime TEXT NOT NULL,
   stock_ticker TEXT NOT NULL,
-  quantity integer NOT NULL,
+  quantity NUMERIC NOT NULL,
   price TEXT NOT NULL,
   status status_type
 );
